pyproject.toml
[project]
name = "nl-arc"
version = "0.1.0"
description = "AI-native machine learning tool that bridges natural language questions and predictive models using declarative Arc-Graph schemas"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Non-linear AI", email = "nonlinearai.org@gmail.com"}
]
maintainers = [
    {name = "Non-linear AI", email = "nonlinearai.org@gmail.com"}
]
requires-python = ">=3.12"
keywords = [
    "machine-learning",
    "ml",
    "ai",
    "pytorch",
    "declarative-ml",
    "arc-graph",
    "llm",
    "natural-language",
    "data-science",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "openai>=1.30.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "duckdb>=1.0.0",
    "pyyaml>=6.0.0",
    "torch>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pluggy>=1.0.0",
    "sqlfluff>=3.0.0",
    "tensorboard>=2.0.0",
    "matplotlib>=3.7.0",
    "pillow>=10.0.0",
    "adbc-driver-snowflake>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/non-linear-ai/arc"
Repository = "https://github.com/non-linear-ai/arc"
Issues = "https://github.com/non-linear-ai/arc/issues"
Documentation = "https://github.com/non-linear-ai/arc/blob/main/README.md"

[project.scripts]
arc = "arc.ui.cli_wrapper:cli_main"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-watcher>=0.4.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["arc*"]

[tool.setuptools.package-data]
arc = [
    "templates/*.yaml",
    "templates/*.j2",
    "core/agents/*/templates/*.j2",
    "core/agents/*/templates/**/*.md",
    "core/agents/shared/examples/*.yaml",
    "resources/**/*.md",
    "resources/**/*.yaml",
    "graph/*.md",
]

[tool.uv]
index-url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG001", "ARG002", "E501"]  # Allow unused args and long lines in tests
"scripts/**/*.py" = ["E501"]  # Allow longer lines in scripts
"src/arc/tools/**/*.py" = ["E501"]  # Allow longer lines in tools (long error messages)
"src/arc/ml/**/*.py" = ["E501"]  # Allow longer lines in ML code (complex expressions)

[tool.ruff.lint.isort]
known-first-party = ["arc"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=arc",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "asyncio: mark test as an asyncio test",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
branch = true
source = ["arc"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
